requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
decorator==4.4.2
discord.py==1.3.1
future==0.18.2
gym==0.17.1
idna==2.9
idna-ssl==1.1.0
imageio==2.8.0
kaggle==1.5.6
kiwisolver==1.1.0
multidict==4.7.5
networkx==2.4
numpy==1.18.1
Pillow==7.0.0
pyglet==1.5.0
pymodm==0.4.2
pymongo==3.10.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-slugify==4.0.0
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
text-unidecode==1.3
tqdm==4.45.0
typing-extensions==3.7.4.1
urllib3==1.24.3
websockets==8.1
yarl==1.4.2
